2002-10-07  Michal Ludvig  <mludvig@suse.cz>
        Merged s390x port by Heiko Carstens <Heiko.Carstens@de.ibm.com>
        and bugfixes to s390 by D.J. Barrow.

+++ stream.c
@@ -1054,7 +1054,7 @@ int arg;
 		break;
 #endif /* SI_GETUDATA */
 	default:
-		printstr(tcp, (int) si.ic_dp, si.ic_len);
+		printstr(tcp, (long) si.ic_dp, si.ic_len);
 		break;
 	}
 	if (exiting(tcp)) {
